Moved Utils.py functions to RestrictionEngine.py
